+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## Unreleased
 
-## 2.0.9 - 2022-02-15
-
 ### Added
 - Added support for PHP 8.1.
 - Added support for more Linux distributions when installing certificates. ([#380](https://github.com/craftcms/nitro/issues/380)) ([#414](https://github.com/craftcms/nitro/issues/414))
+
+### Removed
+- Removed the `blackfire` command.
 
 ## 2.0.8 - 2021-05-18